Refactor: Created common_sections folder and moved first file into it
